Refactor auth imports and error handling in pages
Updated imports to use the new useAuth hook instead of AuthContext in RequireAuth, Login, and Register pages. Improved error handling by using 'unknown' type for caught errors and safer error message extraction in Login, Register, and ResetPassword pages. Also added explicit User type for register API response.
